🚚 rename: admin api 폴더 이동 #63
